requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
Django==3.1.6
django-axes==5.12.0
django-filter==2.4.0
django-ipware==3.0.2
django-letsencrypt==4.0.0
django-oauth-toolkit==1.3.3
django-registration==3.1.1
django-tables2==2.3.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
